Rework validation functions.
The old validation function completely ignored the error message that
lxml provides so *assert_valid* was made to expose that error message
while keeping a clean API.

Both new functions only accept a tree or any of the other options
allowed by lxml's parse function since it was getting too complex to
provide support for accepting 1 or 2 arguments, paths and trees, etc.
This way just accept it just accepts what it can deal with and boots
the rest to lxml.