Add parser recovery for missing/invalid attributes
Recover from missing required attributes by setting an "appropriate"
value for it and produce a warning. Some attributes are absolutely
required (like file paths) so those tags are skipped with a warning.

Recover from invalid attributes (enum-like attributes) by setting it to
either the default or an appropriate one and produce a warning.

Modified the schema so that warnings are not duplicated (very useful for
the validator).

Closes #8